Add online payment

+++ BulochkaWeb/Areas/Customer/Controllers/CartController.cs
@@ -5,6 +5,7 @@ using Bulochka.Utility;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Stripe.Checkout;
 using System.Security.Claims;
 
 namespace BulochkaWeb.Areas.Customer.Controllers
@@ -49,7 +50,8 @@ namespace BulochkaWeb.Areas.Customer.Controllers
             {
                 ListCart = _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value,
                     includeProperties: "Product"),
-                OrderHeader = new()
+                OrderHeader = new(),
+                PaymentOptionList = new SelectList(GetPaymentOptions(), "Value", "Text", 1),
             };
             ShoppingCartVM.OrderHeader.ApplicationUser =
                 _unitOfWork.ApplicationUser.GetFirstOrDefault(u => u.Id == claim.Value);
@@ -74,53 +76,75 @@ namespace BulochkaWeb.Areas.Customer.Controllers
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
 
-            if (claim == null)
-            {
-                throw new ArgumentNullException(nameof(claim));
-            }
-
-            shoppingCartVM.ListCart = _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value,
-                includeProperties: "Product");
-
-            shoppingCartVM.OrderHeader.ApplicationUserId = claim.Value;
-            shoppingCartVM.OrderHeader.ApplicationUser =
-                _unitOfWork.ApplicationUser.GetFirstOrDefault(u => u.Id == claim.Value);
-
+            if (claim == null) throw new ArgumentNullException(nameof(claim));
+            
+            SetListCart(shoppingCartVM, claim);
+            SetUserInfo(shoppingCartVM, claim);
             SetOrderHeaderAddress(shoppingCartVM,
                 shoppingCartVM.OrderHeader.ApplicationUser.City,
                 shoppingCartVM.OrderHeader.ApplicationUser.StreetAddress,
                 shoppingCartVM.OrderHeader.ApplicationUser.PostalCode);
 
-            shoppingCartVM.OrderHeader.OrderStatus = SD.StatusPending;
-            shoppingCartVM.OrderHeader.PaymentStatus = SD.PaymentStatusPending;
-            shoppingCartVM.OrderHeader.OrderDate = DateTime.Now;
+            SetDefaultInformation(shoppingCartVM);
+            CalculateTotal(shoppingCartVM);
 
-            foreach (var cart in shoppingCartVM.ListCart)
+            AddOrderHeaderInfoToDb(shoppingCartVM);
+            AddOrderDetailInfoToDb(shoppingCartVM);
+
+            if (shoppingCartVM.PaymentOption == SD.PaymentOptionOffline)
             {
-                shoppingCartVM.OrderHeader.OrderTotal += (cart.Product.Price * cart.Count);
+                _unitOfWork.ShoppingCart.RemoveRange(shoppingCartVM.ListCart);
+                _unitOfWork.Save();
+                return RedirectToAction("Index", "Home");
             }
-
-            _unitOfWork.OrderHeader.Add(shoppingCartVM.OrderHeader);
-            _unitOfWork.Save();
-
-            foreach (var cart in shoppingCartVM.ListCart)
+            else if (shoppingCartVM.PaymentOption == SD.PaymentOptionOnline)
             {
-                OrderDetail orderDetail = new()
+                var domain = "https://localhost:44324/";
+                var options = new SessionCreateOptions
                 {
-                    ProductId = cart.ProductId,
-                    OrderId = shoppingCartVM.OrderHeader.Id,
-                    Price = cart.Product.Price,
-                    Count = cart.Count,
+                    PaymentMethodTypes = new List<string>
+                    {
+                        "card",
+                    },
+                    LineItems = new List<SessionLineItemOptions>(),
+                    Mode = "payment",
+                    SuccessUrl = domain + $"customer/cart/OrderConfirmation?id={shoppingCartVM.OrderHeader.Id}",
+                    CancelUrl = domain + $"customer/cart/index",
                 };
 
-                _unitOfWork.OrderDetail.Add(orderDetail);
+                foreach (var item in shoppingCartVM.ListCart)
+                {
+                    var sessionLineItem = new SessionLineItemOptions
+                    {
+                        PriceData = new SessionLineItemPriceDataOptions
+                        {
+                            UnitAmount = (long)(item.Product.Price * 100),
+                            Currency = "rub",
+                            ProductData = new SessionLineItemPriceDataProductDataOptions
+                            {
+                                Name = item.Product.Title,
+                            },
+                        },
+                        Quantity = item.Count,
+                    };
+                    options.LineItems.Add(sessionLineItem);
+                }
+
+                var service = new SessionService();
+                Session session = service.Create(options);
+
+                _unitOfWork.OrderHeader.UpdateStripePaymentId(shoppingCartVM.OrderHeader.Id,
+                    session.Id,
+                    session.PaymentIntentId);
                 _unitOfWork.Save();
+
+                Response.Headers.Add("Location", session.Url);
+                return new StatusCodeResult(303);
             }
-
-            _unitOfWork.ShoppingCart.RemoveRange(shoppingCartVM.ListCart);
-            _unitOfWork.Save();
-
-            return RedirectToAction("Index", "Home");
+            else
+            {
+                throw new InvalidDataException("invalid payment option");
+            }
         }
 
         public IActionResult SummaryInCafe()
@@ -138,7 +162,8 @@ namespace BulochkaWeb.Areas.Customer.Controllers
                     {
                         Text = b.City + ", " + b.StreetAddress,
                         Value = b.Id.ToString()
-                    })
+                    }),
+                PaymentOptionList = new SelectList(GetPaymentOptions(), "Value", "Text", 1),
             };
             ShoppingCartVM.OrderHeader.ApplicationUser =
                 _unitOfWork.ApplicationUser.GetFirstOrDefault(u => u.Id == claim.Value);
@@ -159,32 +184,130 @@ namespace BulochkaWeb.Areas.Customer.Controllers
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
 
-            shoppingCartVM.ListCart = _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value,
-                includeProperties: "Product");
+            if (claim == null) throw new ArgumentNullException(nameof(claim));
 
             var companyBranch = _unitOfWork.CompanyBranch.GetFirstOrDefault(
                 b => b.Id == shoppingCartVM.OrderHeader.PickUpPlaceId);
-            
-            SetOrderHeaderAddress(shoppingCartVM, 
-                companyBranch.City, 
-                companyBranch.StreetAddress, 
+
+            SetListCart(shoppingCartVM, claim);
+            SetUserInfo(shoppingCartVM, claim);
+            SetOrderHeaderAddress(shoppingCartVM,
+                companyBranch.City,
+                companyBranch.StreetAddress,
                 companyBranch.PostalCode);
 
+            SetDefaultInformation(shoppingCartVM);
+            CalculateTotal(shoppingCartVM);
+
+            AddOrderHeaderInfoToDb(shoppingCartVM);
+            AddOrderDetailInfoToDb(shoppingCartVM);
+
+            if (shoppingCartVM.PaymentOption == SD.PaymentOptionOffline)
+            {
+                _unitOfWork.ShoppingCart.RemoveRange(shoppingCartVM.ListCart);
+                _unitOfWork.Save();
+                return RedirectToAction("Index", "Home");
+            }
+            else if (shoppingCartVM.PaymentOption == SD.PaymentOptionOnline)
+            {
+                var domain = "https://localhost:44324/";
+                var options = new SessionCreateOptions
+                {
+                    PaymentMethodTypes = new List<string> { "card" },
+                    LineItems = new List<SessionLineItemOptions>(),
+                    Mode = "payment",
+                    SuccessUrl = domain + $"customer/cart/OrderConfirmation?id={shoppingCartVM.OrderHeader.Id}",
+                    CancelUrl = domain + $"customer/cart/index",
+                };
+
+                foreach (var item in shoppingCartVM.ListCart)
+                {
+                    var sessionLineItem = new SessionLineItemOptions
+                    {
+                        PriceData = new SessionLineItemPriceDataOptions
+                        {
+                            UnitAmount = (long)(item.Product.Price * 100),
+                            Currency = "rub",
+                            ProductData = new SessionLineItemPriceDataProductDataOptions
+                            {
+                                Name = item.Product.Title,
+                            },
+                        },
+                        Quantity = item.Count,
+                    };
+                    options.LineItems.Add(sessionLineItem);
+                }
+
+                var service = new SessionService();
+                Session session = service.Create(options);
+
+                _unitOfWork.OrderHeader.UpdateStripePaymentId(shoppingCartVM.OrderHeader.Id,
+                    session.Id,
+                    session.PaymentIntentId);
+                _unitOfWork.Save();
+
+                Response.Headers.Add("Location", session.Url);
+                return new StatusCodeResult(303); 
+            }
+            else
+            {
+                throw new InvalidDataException("invalid payment option");
+            }
+        }
+
+        public IActionResult OrderConfirmation(int id)
+        {
+            OrderHeader orderHeader = _unitOfWork.OrderHeader.GetFirstOrDefault(u => u.Id == id);
+
+            SessionService service = new SessionService();
+            Session session = service.Get(orderHeader.SessionId);
+            if (session.PaymentStatus.ToLower() == SD.Paid)
+            {
+                _unitOfWork.OrderHeader.UpdateStatus(id, SD.StatusApproved, SD.PaymentStatusApproved);
+                _unitOfWork.Save();
+            }
+
+            List<ShoppingCart> shoppingCarts = _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == orderHeader.ApplicationUserId).ToList();
+            _unitOfWork.ShoppingCart.RemoveRange(shoppingCarts);
+            _unitOfWork.Save();
+            return View();
+        }
+
+        private List<SelectListItem> GetPaymentOptions()
+        {
+            return new List<SelectListItem>
+            {
+                new SelectListItem
+                {
+                    Text = "Оплатить при получении",
+                    Value = SD.PaymentOptionOffline,
+                },
+                new SelectListItem
+                {
+                    Text = "Оплатить на сайте",
+                    Value = SD.PaymentOptionOnline,
+                },
+            };
+        }
+
+        #region SummaryPOST common methods
+
+        private static void SetOrderHeaderAddress(ShoppingCartVM ShoppingCartVM, string city, string address, string postalCode)
+        {
+            ShoppingCartVM.OrderHeader.City = city;
+            ShoppingCartVM.OrderHeader.StreetAddress = address;
+            ShoppingCartVM.OrderHeader.PostalCode = postalCode;
+        }
+
+        private static void SetDefaultInformation(ShoppingCartVM shoppingCartVM)
+        {
             shoppingCartVM.OrderHeader.OrderStatus = SD.StatusPending;
             shoppingCartVM.OrderHeader.PaymentStatus = SD.PaymentStatusPending;
             shoppingCartVM.OrderHeader.OrderDate = DateTime.Now;
-            shoppingCartVM.OrderHeader.ApplicationUserId = claim.Value;
-            shoppingCartVM.OrderHeader.ApplicationUser =
-                _unitOfWork.ApplicationUser.GetFirstOrDefault(u => u.Id == claim.Value);
+        }
 
-            foreach (var cart in shoppingCartVM.ListCart)
-            {
-                shoppingCartVM.OrderHeader.OrderTotal += (cart.Product.Price * cart.Count);
-            }
-
-            _unitOfWork.OrderHeader.Add(shoppingCartVM.OrderHeader);
-            _unitOfWork.Save();
-
+        private void AddOrderDetailInfoToDb(ShoppingCartVM shoppingCartVM)
+        {
             foreach (var cart in shoppingCartVM.ListCart)
             {
                 OrderDetail orderDetail = new()
@@ -198,19 +321,38 @@ namespace BulochkaWeb.Areas.Customer.Controllers
                 _unitOfWork.OrderDetail.Add(orderDetail);
                 _unitOfWork.Save();
             }
-
-            _unitOfWork.ShoppingCart.RemoveRange(shoppingCartVM.ListCart);
-            _unitOfWork.Save();
-
-            return RedirectToAction("Index", "Home");
         }
 
-        private static void SetOrderHeaderAddress(ShoppingCartVM ShoppingCartVM, string city, string address, string postalCode)
+        private static void CalculateTotal(ShoppingCartVM shoppingCartVM)
         {
-            ShoppingCartVM.OrderHeader.City = city;
-            ShoppingCartVM.OrderHeader.StreetAddress = address;
-            ShoppingCartVM.OrderHeader.PostalCode = postalCode;
+            foreach (var cart in shoppingCartVM.ListCart)
+            {
+                shoppingCartVM.OrderHeader.OrderTotal += (cart.Product.Price * cart.Count);
+            }
         }
+
+        private void SetUserInfo(ShoppingCartVM shoppingCartVM, Claim? claim)
+        {
+            shoppingCartVM.OrderHeader.ApplicationUserId = claim.Value;
+            shoppingCartVM.OrderHeader.ApplicationUser =
+                _unitOfWork.ApplicationUser.GetFirstOrDefault(u => u.Id == claim.Value);
+        }
+
+        private void AddOrderHeaderInfoToDb(ShoppingCartVM shoppingCartVM)
+        {
+            _unitOfWork.OrderHeader.Add(shoppingCartVM.OrderHeader);
+            _unitOfWork.Save();
+        }
+
+        private void SetListCart(ShoppingCartVM shoppingCartVM, Claim? claim)
+        {
+            shoppingCartVM.ListCart = _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value,
+                includeProperties: "Product");
+        }
+
+        #endregion
+
+        #region Cart methods
 
         public IActionResult Plus(int cartId)
         {
@@ -242,6 +384,9 @@ namespace BulochkaWeb.Areas.Customer.Controllers
             _unitOfWork.ShoppingCart.Remove(cart);
             _unitOfWork.Save();
             return RedirectToAction(nameof(Index));
-        }
+        } 
+
+        #endregion
+
     }
 }
